Added a new property to the redux state: sortButtonState.
The sortButtonState has a custom type which is an object literal with two properties, the category property and the sort property.

The category property can be one of four different strings, representing the different categories to sort by.

The sort property can be one of two different strings, ascending or descending, to represent which sort order to use.

A reducer has been added for updating the sortButtonState property on the redux state.

This reducer is dispatched by the SortButton components, which contain buttons that the user can click on to sort the data as desired.  The Viewer components watch for changes to the sortButtonState via useEffect hooks, and will sort the data according to the property valies in the sortButtonState.